Support latest service-catalog in OS 3.7 (#131)
* Support latest service-catalog in OS 3.7

Update the ansible service broker setup to wait for the service-catalog
to be available. This prevents a race condition when we are working with
coalmine and the service-catalog (giving it time to redeploy).

Also, with the service-catalog being pushed into `kube-service-catalog`
update the openshift_setup to register the project where the service
catalog was put.

* Service broker setup should handle new template

The new broker template includes the broker_kind and broker_auth
variables that should be overridable, this makes it possible to override
those values.

* Give service catalog more permissions

When using the latest service catalog, it will need cluser-admin role in
order to function properly. This update gives the controller service
account for the service-catalog cluster-admin role.

* Use http with latest service catalog

Update to allow latest service catalog to work with OS 3.7 by rendering
the template with user provided values for endpoint termination, etc.

Update the example for coalmine as you only need to specify the images
in the case that you want to override the defaults.

* Simplify on/off of broker ssl settings

Who really wants to set 5 variables to turn ssl off when 1 would do? No
one, so this update makes it so the user just has to set broker_ssl to
false and the broker will be created without ssl settings, voila.