Kubernetes playbook (#120)
* Create the Kubernetes playbook

Create a simple skeleton playbook and reformat the existing
playbook so Kubernetes can fit in at the top level.

* Setup scripts to move around cluster vars

The Cluster variable will be passed around similar to hostname.

* Add the kubeadm ansible role

* Fill out the kubernetes playbook

This get the kubernetes playbook working, but doesn't integrate
with the broker playbook yet.

* Spawn the service catalog

Create the service catalog using helm charts.

* Use the canary tag with creating images

* Spawn the Broker on Kubernetes

Create the Broker resources on Kubernetes

* Change docker_org_name to docker_org

Use docker_org as a var name since it matches the one in the template

* Few changes in the templates

Fix the namespace switch for kubernetes

* Install the cfssl pkg

The cfssl go pkg is needed to generate ssl certs

* Fix the catalog's connection to the api server

The catalog expected a cert in order to identify with the
api server.

* Ignore ssl certs generated from service-catalog script

* Add another wait for the tiller server

* Add iptables commands to the playbook

We need to allow traffic to reach outside of the pod network.

* Change the broker route to an endpoint

Catalog will contact the broker using its endpoint

* Few more fixes to the initial setup tasks

* Change the complete messsage for Kubernetes

* Build the latest catalog images each run

Since we pull in the newest catalog each run, we also need
the latest binaries to match.  Build the catalog containers
each run.

* Remove the catalog repo each commit

* Fix up a bunch of permission issues and add a validation

The kubectl client needs to be >= 1.7. Let's validate that
during the playbook.

* Make sure we're running the locally build catalog images

Before, the catalog images we were building weren't being run.
Also, use helm's cli to set variables instead of using regexes.

* Change the iptables rule based on the docker0 address